requirements-dev.txt
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-flask>=1.2.0

ruff>=0.6.0
black>=24.0.0
mypy>=1.11.0
bandit>=1.7.0
pip-audit>=2.7.0

pre-commit>=3.5.0
